stage 5 complete

+++ src/main/java/com/mihey/quiz/QuizController.java
@@ -5,7 +5,10 @@ import com.mihey.quiz.model_quiz.Quiz;
 import com.mihey.quiz.model_user.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -35,6 +38,10 @@ public class QuizController {
 
     @PostMapping("/api/quizzes")
     public Quiz addQuiz(@Valid @RequestBody Quiz quiz) {
+        SecurityContext context = SecurityContextHolder.getContext();
+        Authentication authentication = context.getAuthentication();
+        String username = authentication.getName();
+        quiz.setEmail(username);
         return quizRepository.save(quiz);
     }
 
@@ -64,12 +71,18 @@ public class QuizController {
     }
 
     @DeleteMapping("/api/quizzes/{id}")
-    public void deleteQuizById(@AuthenticationPrincipal User user, @PathVariable long id) {
+    public void deleteQuizById(@PathVariable long id) {
+        SecurityContext context = SecurityContextHolder.getContext();
+        Authentication authentication = context.getAuthentication();
+        String username = authentication.getName();
         if (id < 1 || quizRepository.findById(id).isEmpty()) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
-        } else {
+        }
+        if (!quizRepository.findById(id).get().getEmail().equals(username)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
             quizRepository.deleteById(id);
             throw new ResponseStatusException(HttpStatus.NO_CONTENT);
-        }
+
     }
 }
